+++ common/src/main/java/com/lowdragmc/lowdraglib/client/renderer/impl/UIResourceRenderer.java
@@ -3,6 +3,7 @@ package com.lowdragmc.lowdraglib.client.renderer.impl;
 import com.lowdragmc.lowdraglib.client.renderer.IRenderer;
 import com.lowdragmc.lowdraglib.gui.editor.data.resource.Resource;
 import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.datafixers.util.Either;
 import lombok.Getter;
 import lombok.Setter;
 import net.fabricmc.api.EnvType;
@@ -24,6 +25,7 @@ import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.io.File;
 import java.util.List;
 import java.util.function.Consumer;
 
@@ -46,13 +48,13 @@ public class UIResourceRenderer implements IRenderer {
     @Setter
     private Resource<IRenderer> resource;
 
-    public final String key;
+    public final Either<String, File> key;
 
-    public UIResourceRenderer(String key) {
+    public UIResourceRenderer(Either<String, File> key) {
         this.key = key;
     }
 
-    public UIResourceRenderer(Resource<IRenderer> resource, String key) {
+    public UIResourceRenderer(Resource<IRenderer> resource, Either<String, File> key) {
         this.resource = resource;
         this.key = key;
     }