+++ common/src/main/java/com/lowdragmc/lowdraglib/gui/editor/data/Resources.java
@@ -1,6 +1,9 @@
 package com.lowdragmc.lowdraglib.gui.editor.data;
 
+import com.lowdragmc.lowdraglib.gui.editor.data.resource.ColorsResource;
+import com.lowdragmc.lowdraglib.gui.editor.data.resource.EntriesResource;
 import com.lowdragmc.lowdraglib.gui.editor.data.resource.Resource;
+import com.lowdragmc.lowdraglib.gui.editor.data.resource.TexturesResource;
 import com.lowdragmc.lowdraglib.gui.editor.runtime.AnnotationDetector;
 import net.minecraft.nbt.CompoundTag;
 
@@ -16,14 +19,6 @@ public class Resources {
 
     public final Map<String, Resource<?>> resources;
 
-    protected Resources() {
-        this.resources = new LinkedHashMap<>();
-        for (var wrapper : AnnotationDetector.REGISTER_RESOURCES) {
-            var resource =  wrapper.creator().get();
-            resources.put(resource.name(), resource);
-        }
-    }
-
     public Resources(Map<String, Resource<?>> resources) {
         this.resources = resources;
     }
@@ -33,15 +28,32 @@ public class Resources {
     }
 
     public static Resources fromNBT(CompoundTag tag) {
-        var resource = new Resources();
-        resource.deserializeNBT(tag);
-        return resource;
+        Map<String, Resource<?>> map = new LinkedHashMap<>();
+        for (String key : tag.getAllKeys()) {
+            var resource = AnnotationDetector.REGISTER_RESOURCES.stream()
+                    .filter(wrapper -> wrapper.annotation().name().equals(key))
+                    .findFirst();
+            if (resource.isEmpty()) continue;
+            map.put(key, resource.get().creator().get());
+        }
+        var resources = new Resources(map);
+        resources.deserializeNBT(tag);
+        return resources;
     }
 
+    @Deprecated(since = "1.21")
     public static Resources defaultResource() { // default
-        Resources resources = new Resources();
+        var resources = of(new EntriesResource(), new ColorsResource(), new TexturesResource());
         resources.resources.values().forEach(Resource::buildDefault);
         return resources;
+    }
+
+    public static Resources of(Resource<?>... resources) { // default
+        Map<String, Resource<?>> map = new LinkedHashMap<>();
+        for (Resource<?> resource : resources) {
+            map.put(resource.name(), resource);
+        }
+        return new Resources(map);
     }
 
     @SuppressWarnings({"rawtypes", "unchecked"})