+++ common/src/main/java/com/lowdragmc/lowdraglib/syncdata/accessor/IGuiTextureAccessor.java
@@ -5,7 +5,10 @@ import com.lowdragmc.lowdraglib.gui.texture.UIResourceTexture;
 import com.lowdragmc.lowdraglib.syncdata.AccessorOp;
 import com.lowdragmc.lowdraglib.syncdata.payload.ITypedPayload;
 import com.lowdragmc.lowdraglib.syncdata.payload.NbtTagPayload;
+import com.mojang.datafixers.util.Either;
 import net.minecraft.nbt.CompoundTag;
+
+import java.io.File;
 
 /**
  * @author KilaBash
@@ -25,7 +28,19 @@ public class IGuiTextureAccessor extends CustomObjectAccessor<IGuiTexture>{
             tag = new CompoundTag();
             if (value instanceof UIResourceTexture uiResourceTexture) {
                 tag.putString("type", "ui_resource");
-                tag.putString("key", uiResourceTexture.key);
+                tag.put("key", uiResourceTexture.key.map(
+                        l -> {
+                            var key = new CompoundTag();
+                            key.putString("key", l);
+                            key.putString("type", "builtin");
+                            return key;
+                        }, r-> {
+                            var key = new CompoundTag();
+                            key.putString("key", r.getPath());
+                            key.putString("type", "project");
+                            return key;
+                        }
+                ));
             } else {
                 tag.putString("type", "empty");
             }
@@ -39,7 +54,19 @@ public class IGuiTextureAccessor extends CustomObjectAccessor<IGuiTexture>{
             var tag = (CompoundTag)nbtTagPayload.getPayload();
             var type = tag.getString("type");
             if (type.equals("ui_resource") && tag.contains("key")) {
-                var key = tag.getString("key");
+                var keyTag = tag.get("key");
+                Either<String, File> key = null;
+                if (keyTag instanceof CompoundTag compoundTag) {
+                    var keyType = compoundTag.getString("type");
+                    var keyValue = compoundTag.getString("key");
+                    if (keyType.equals("builtin")) {
+                        key = Either.left(keyValue);
+                    } else if (keyType.equals("project")) {
+                        key = Either.right(new File(keyValue));
+                    }
+                } else if (keyTag != null){
+                    key = Either.left(keyTag.getAsString());
+                }
                 var resource = UIResourceTexture.getProjectResource();
                 if (resource == null) {
                     return new UIResourceTexture(key);
@@ -47,7 +74,7 @@ public class IGuiTextureAccessor extends CustomObjectAccessor<IGuiTexture>{
                 if (UIResourceTexture.isProject()) {
                     return new UIResourceTexture(resource, key);
                 } else {
-                    return resource.getResourceOrDefault(key, IGuiTexture.MISSING_TEXTURE);
+                    return key == null ? IGuiTexture.MISSING_TEXTURE: resource.getResourceOrDefault(key, IGuiTexture.MISSING_TEXTURE);
                 }
             }
             return IGuiTexture.deserializeWrapper(tag);