+++ common/src/main/java/com/lowdragmc/lowdraglib/utils/LocalizationUtils.java
@@ -34,8 +34,8 @@ public class LocalizationUtils {
         if (!LDLib.isClient()) {
             return String.format(localisationKey, substitutions);
         } else {
-            if (RESOURCE != null && RESOURCE.hasResource(localisationKey)) {
-                return RESOURCE.getResource(localisationKey);
+            if (RESOURCE != null && RESOURCE.hasBuiltinResource(localisationKey)) {
+                return RESOURCE.getBuiltinResource(localisationKey);
             }
             return I18n.get(localisationKey, substitutions);
         }