+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 * Fixed field synchronized in multiple threads
 * Qol double click to open the file in the editor
+* Qol added static resource supports
 * Added more APIs for Transform